+++ src/main/java/ru/clevertec/writer/Writer.java
@@ -1,5 +1,7 @@
 package ru.clevertec.writer;
 
+import org.aspectj.lang.annotation.After;
+
 /**
  * Класс для вызова соответствующего класса,
  * осуществляющего запись в файл
