+++ src/main/java/ru/clevertec/writer/impl/WriterPdf.java
@@ -11,6 +11,7 @@ import ru.clevertec.writer.util.pdf.impl.CreatePdfPageWithTemplateImpl;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Objects;
 
@@ -39,7 +40,7 @@ public class WriterPdf<T> implements IWriter<T> {
     @Override
     public void createFile(String caption, T entity) {
         String dest = DEST_PATH + entity.getClass().getSimpleName() + "_"
-                + LocalDate.now().format(DateTimeFormatter.ISO_DATE) + ".pdf";
+                + LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd_HH-mm-ss")) + ".pdf";
         PdfDocument destPdf = createPdfWriter(dest);
         if (FILE_TEMPLATE.isEmpty()) {
             new CreatePdfPageImpl(destPdf).createPage();
