+++ src/test/java/ru/clevertec/service/impl/AnimalServiceImplTest.java
@@ -1,5 +1,6 @@
-package ru.clevertec.animal.service.impl;
+package ru.clevertec.service.impl;
 
+import net.bytebuddy.implementation.bind.annotation.IgnoreForBinding;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -22,6 +23,7 @@ import java.util.UUID;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.verify;
+
 
 @ExtendWith(MockitoExtension.class)
 class AnimalServiceImplTest {
